:refactor: Map & presentationControllerDidDismiss refactoring
- CommonUI에서 Map 관련 폴더링 및 구조 개선
- MyLocationViewController에서 CommonUI의 MapViewable 적용
- presentationControllerDidDismiss 메소드에서 기존 PlaceSearch Riblet을 detach하는 로직에서 modal로 띄워진 riblet들 detach하는 메소드로 로직 수정 및 메소드명 변경
