+++ PlaceStory/PlaceDiary/Package.swift
@@ -12,11 +12,18 @@ let package = Package(
             name: "PlaceList",
             targets: ["PlaceList"]),
     ],
+    dependencies: [
+        .package(path: "../ProxyPackage")
+    ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
         // Targets can depend on other targets in this package and products from dependencies.
         .target(
-            name: "PlaceList"),
+            name: "PlaceList",
+            dependencies: [
+                "ProxyPackage"
+            ]
+        ),
         .testTarget(
             name: "PlaceDiaryTests",
             dependencies: ["PlaceList"]),
